Remove link between bblfsh/web and bblfsh/bblfshd
The link is not necessary, maybe it was for some previous
version. As a side-effect that link was adding an additional name to
the `srcd-cli-bblfshd` container that sometimes prevents it from being
pruned correctly.

The problem occurred because during the pruning we use the
first name in the list of names of a container. Once `srcd-cli-bblfshd-web` is
started, the first name of the bblfsh server becomes
`srcd-cli-bblfsh-web/srcd-cli-bblfshd` (due to the link explained
above). This "extra" name is removed once the container
`srcd-cli-bblfshd-web` is removed. So when listing the containers to prune, if
`srcd-cli-bblfshd-web` is removed before
`srcd-cli-bblfsh-web/srcd-cli-bblfshd`,
when we try to remove `srcd-cli-bblfsh-web/srcd-cli-bblfshd` that
container doesn't actually exist anymore, at least it doesn't exist
with that name.

Signed-off-by: Lou Marvin Caraig <loumarvincaraig@gmail.com>